refactor(views): Dashboard, Slice, Database, SQL_Lab into separate folders
