+++ superset/views/Dashboard/__init__.py
@@ -1,0 +1,2 @@
+from .dashboard import DashboardModelView
+from .dashboard import DashboardModelViewAsync
