+++ superset/views/Database/__init__.py
@@ -1,0 +1,3 @@
+from .database import DatabaseView
+from .database import DatabaseAsync
+from .database import DatabaseTablesAsync
