+++ superset/views/SQL_Lab/__init__.py
@@ -1,0 +1,4 @@
+from .sql_lab import QueryView
+from .sql_lab import SavedQueryView
+from .sql_lab import SavedQueryViewApi
+from .sql_lab import SqlLab
