+++ superset/views/Slice/__init__.py
@@ -1,0 +1,4 @@
+from .slice import SliceFilter
+from .slice import SliceModelView
+from .slice import SliceAsync
+from .slice import SliceAddView
