+++ superset/views/__init__.py
@@ -1,3 +1,6 @@
 from . import base  # noqa
 from . import core  # noqa
+from . import database  # noqa
+from . import slice  # noqa
+from . import dashboard  # noqa
 from . import sql_lab  # noqa
