+++ superset/views/core.py
@@ -120,42 +120,6 @@ def check_ownership(obj, raise_if_false=True):
         return False
 
 
-class SliceFilter(SupersetFilter):
-    def apply(self, query, func):  # noqa
-        if self.has_all_datasource_access():
-            return query
-        perms = self.get_view_menus('datasource_access')
-        # TODO(bogdan): add `schema_access` support here
-        return query.filter(self.model.perm.in_(perms))
-
-
-class DashboardFilter(SupersetFilter):
-
-    """List dashboards for which users have access to at least one slice"""
-
-    def apply(self, query, func):  # noqa
-        if self.has_all_datasource_access():
-            return query
-        Slice = models.Slice  # noqa
-        Dash = models.Dashboard  # noqa
-        # TODO(bogdan): add `schema_access` support here
-        datasource_perms = self.get_view_menus('datasource_access')
-        slice_ids_qry = (
-            db.session
-            .query(Slice.id)
-            .filter(Slice.perm.in_(datasource_perms))
-        )
-        query = query.filter(
-            Dash.id.in_(
-                db.session.query(Dash.id)
-                .distinct()
-                .join(Dash.slices)
-                .filter(Slice.id.in_(slice_ids_qry))
-            )
-        )
-        return query
-
-
 def generate_download_headers(extension):
     filename = datetime.now().strftime("%Y%m%d_%H%M%S")
     content_disp = "attachment; filename={}.{}".format(filename, extension)
@@ -163,126 +127,6 @@ def generate_download_headers(extension):
         "Content-Disposition": content_disp,
     }
     return headers
-
-
-class DatabaseView(SupersetModelView, DeleteMixin):  # noqa
-    datamodel = SQLAInterface(models.Database)
-    list_columns = [
-        'database_name', 'backend', 'allow_run_sync', 'allow_run_async',
-        'allow_dml', 'creator', 'modified']
-    add_columns = [
-        'database_name', 'sqlalchemy_uri', 'cache_timeout', 'extra',
-        'expose_in_sqllab', 'allow_run_sync', 'allow_run_async',
-        'allow_ctas', 'allow_dml', 'force_ctas_schema']
-    search_exclude_columns = ('password',)
-    edit_columns = add_columns
-    show_columns = [
-        'tables',
-        'cache_timeout',
-        'extra',
-        'database_name',
-        'sqlalchemy_uri',
-        'perm',
-        'created_by',
-        'created_on',
-        'changed_by',
-        'changed_on',
-    ]
-    add_template = "superset/models/database/add.html"
-    edit_template = "superset/models/database/edit.html"
-    base_order = ('changed_on', 'desc')
-    description_columns = {
-        'sqlalchemy_uri': utils.markdown(
-            "Refer to the "
-            "[SqlAlchemy docs]"
-            "(http://docs.sqlalchemy.org/en/rel_1_0/core/engines.html#"
-            "database-urls) "
-            "for more information on how to structure your URI.", True),
-        'expose_in_sqllab': _("Expose this DB in SQL Lab"),
-        'allow_run_sync': _(
-            "Allow users to run synchronous queries, this is the default "
-            "and should work well for queries that can be executed "
-            "within a web request scope (<~1 minute)"),
-        'allow_run_async': _(
-            "Allow users to run queries, against an async backend. "
-            "This assumes that you have a Celery worker setup as well "
-            "as a results backend."),
-        'allow_ctas': _("Allow CREATE TABLE AS option in SQL Lab"),
-        'allow_dml': _(
-            "Allow users to run non-SELECT statements "
-            "(UPDATE, DELETE, CREATE, ...) "
-            "in SQL Lab"),
-        'force_ctas_schema': _(
-            "When allowing CREATE TABLE AS option in SQL Lab, "
-            "this option forces the table to be created in this schema"),
-        'extra': utils.markdown(
-            "JSON string containing extra configuration elements. "
-            "The ``engine_params`` object gets unpacked into the "
-            "[sqlalchemy.create_engine]"
-            "(http://docs.sqlalchemy.org/en/latest/core/engines.html#"
-            "sqlalchemy.create_engine) call, while the ``metadata_params`` "
-            "gets unpacked into the [sqlalchemy.MetaData]"
-            "(http://docs.sqlalchemy.org/en/rel_1_0/core/metadata.html"
-            "#sqlalchemy.schema.MetaData) call. ", True),
-    }
-    label_columns = {
-        'expose_in_sqllab': _("Expose in SQL Lab"),
-        'allow_ctas': _("Allow CREATE TABLE AS"),
-        'allow_dml': _("Allow DML"),
-        'force_ctas_schema': _("CTAS Schema"),
-        'database_name': _("Database"),
-        'creator': _("Creator"),
-        'changed_on_': _("Last Changed"),
-        'sqlalchemy_uri': _("SQLAlchemy URI"),
-        'cache_timeout': _("Cache Timeout"),
-        'extra': _("Extra"),
-    }
-
-    def pre_add(self, db):
-        db.set_sqlalchemy_uri(db.sqlalchemy_uri)
-        security.merge_perm(sm, 'database_access', db.perm)
-        for schema in db.all_schema_names():
-            security.merge_perm(
-                sm, 'schema_access', utils.get_schema_perm(db, schema))
-
-    def pre_update(self, db):
-        self.pre_add(db)
-
-
-appbuilder.add_link(
-    'Import Dashboards',
-    label=__("Import Dashboards"),
-    href='/superset/import_dashboards',
-    icon="fa-cloud-upload",
-    category='Manage',
-    category_label=__("Manage"),
-    category_icon='fa-wrench',)
-
-
-appbuilder.add_view(
-    DatabaseView,
-    "Databases",
-    label=__("Databases"),
-    icon="fa-database",
-    category="Sources",
-    category_label=__("Sources"),
-    category_icon='fa-database',)
-
-
-class DatabaseAsync(DatabaseView):
-    list_columns = [
-        'id', 'database_name',
-        'expose_in_sqllab', 'allow_ctas', 'force_ctas_schema',
-        'allow_run_async', 'allow_run_sync', 'allow_dml',
-    ]
-
-appbuilder.add_view_no_menu(DatabaseAsync)
-
-
-class DatabaseTablesAsync(DatabaseView):
-    list_columns = ['id', 'all_table_names', 'all_schema_names']
-
-appbuilder.add_view_no_menu(DatabaseTablesAsync)
 
 
 class AccessRequestsModelView(SupersetModelView, DeleteMixin):
@@ -308,200 +152,6 @@ appbuilder.add_view(
     category="Security",
     category_label=__("Security"),
     icon='fa-table',)
-
-
-class SliceModelView(SupersetModelView, DeleteMixin):  # noqa
-    datamodel = SQLAInterface(models.Slice)
-    can_add = False
-    label_columns = {
-        'datasource_link': 'Datasource',
-    }
-    list_columns = [
-        'slice_link', 'viz_type', 'datasource_link', 'creator', 'modified']
-    edit_columns = [
-        'slice_name', 'description', 'viz_type', 'owners', 'dashboards',
-        'params', 'cache_timeout']
-    base_order = ('changed_on', 'desc')
-    description_columns = {
-        'description': Markup(
-            "The content here can be displayed as widget headers in the "
-            "dashboard view. Supports "
-            "<a href='https://daringfireball.net/projects/markdown/'>"
-            "markdown</a>"),
-        'params': _(
-            "These parameters are generated dynamically when clicking "
-            "the save or overwrite button in the explore view. This JSON "
-            "object is exposed here for reference and for power users who may "
-            "want to alter specific parameters."),
-        'cache_timeout': _(
-            "Duration (in seconds) of the caching timeout for this slice."
-        ),
-    }
-    base_filters = [['id', SliceFilter, lambda: []]]
-    label_columns = {
-        'cache_timeout': _("Cache Timeout"),
-        'creator': _("Creator"),
-        'dashboards': _("Dashboards"),
-        'datasource_link': _("Datasource"),
-        'description': _("Description"),
-        'modified': _("Last Modified"),
-        'owners': _("Owners"),
-        'params': _("Parameters"),
-        'slice_link': _("Slice"),
-        'slice_name': _("Name"),
-        'table': _("Table"),
-        'viz_type': _("Visualization Type"),
-    }
-
-    def pre_update(self, obj):
-        check_ownership(obj)
-
-    def pre_delete(self, obj):
-        check_ownership(obj)
-
-    @expose('/add', methods=['GET', 'POST'])
-    @has_access
-    def add(self):
-        flash(__(
-            "To create a new slice, you can open a data source "
-            "through the `Sources` menu, or alter an existing slice "
-            "from the `Slices` menu"), "info")
-        return redirect('/superset/welcome')
-
-appbuilder.add_view(
-    SliceModelView,
-    "Slices",
-    label=__("Slices"),
-    icon="fa-bar-chart",
-    category="",
-    category_icon='',)
-
-
-class SliceAsync(SliceModelView):  # noqa
-    list_columns = [
-        'slice_link', 'viz_type',
-        'creator', 'modified', 'icons']
-    label_columns = {
-        'icons': ' ',
-        'slice_link': _('Slice'),
-    }
-
-appbuilder.add_view_no_menu(SliceAsync)
-
-
-class SliceAddView(SliceModelView):  # noqa
-    list_columns = [
-        'id', 'slice_name', 'slice_link', 'viz_type',
-        'owners', 'modified', 'changed_on']
-
-appbuilder.add_view_no_menu(SliceAddView)
-
-
-class DashboardModelView(SupersetModelView, DeleteMixin):  # noqa
-    datamodel = SQLAInterface(models.Dashboard)
-    list_columns = ['dashboard_link', 'creator', 'modified']
-    edit_columns = [
-        'dashboard_title', 'slug', 'slices', 'owners', 'position_json', 'css',
-        'json_metadata']
-    show_columns = edit_columns + ['table_names']
-    add_columns = edit_columns
-    base_order = ('changed_on', 'desc')
-    description_columns = {
-        'position_json': _(
-            "This json object describes the positioning of the widgets in "
-            "the dashboard. It is dynamically generated when adjusting "
-            "the widgets size and positions by using drag & drop in "
-            "the dashboard view"),
-        'css': _(
-            "The css for individual dashboards can be altered here, or "
-            "in the dashboard view where changes are immediately "
-            "visible"),
-        'slug': _("To get a readable URL for your dashboard"),
-        'json_metadata': _(
-            "This JSON object is generated dynamically when clicking "
-            "the save or overwrite button in the dashboard view. It "
-            "is exposed here for reference and for power users who may "
-            "want to alter specific parameters."),
-        'owners': _("Owners is a list of users who can alter the dashboard."),
-    }
-    base_filters = [['slice', DashboardFilter, lambda: []]]
-    add_form_query_rel_fields = {
-        'slices': [['slices', SliceFilter, None]],
-    }
-    edit_form_query_rel_fields = add_form_query_rel_fields
-    label_columns = {
-        'dashboard_link': _("Dashboard"),
-        'dashboard_title': _("Title"),
-        'slug': _("Slug"),
-        'slices': _("Slices"),
-        'owners': _("Owners"),
-        'creator': _("Creator"),
-        'modified': _("Modified"),
-        'position_json': _("Position JSON"),
-        'css': _("CSS"),
-        'json_metadata': _("JSON Metadata"),
-        'table_names': _("Underlying Tables"),
-    }
-
-    def pre_add(self, obj):
-        obj.slug = obj.slug.strip() or None
-        if obj.slug:
-            obj.slug = obj.slug.replace(" ", "-")
-            obj.slug = re.sub(r'\W+', '', obj.slug)
-        if g.user not in obj.owners:
-            obj.owners.append(g.user)
-        utils.validate_json(obj.json_metadata)
-        utils.validate_json(obj.position_json)
-        owners = [o for o in obj.owners]
-        for slc in obj.slices:
-            slc.owners = list(set(owners) | set(slc.owners))
-
-    def pre_update(self, obj):
-        check_ownership(obj)
-        self.pre_add(obj)
-
-    def pre_delete(self, obj):
-        check_ownership(obj)
-
-    @action("mulexport", __("Export"), __("Export dashboards?"), "fa-database")
-    def mulexport(self, items):
-        ids = ''.join('&id={}'.format(d.id) for d in items)
-        return redirect(
-            '/dashboardmodelview/export_dashboards_form?{}'.format(ids[1:]))
-
-    @expose("/export_dashboards_form")
-    def download_dashboards(self):
-        if request.args.get('action') == 'go':
-            ids = request.args.getlist('id')
-            return Response(
-                models.Dashboard.export_dashboards(ids),
-                headers=generate_download_headers("pickle"),
-                mimetype="application/text")
-        return self.render_template(
-            'superset/export_dashboards.html',
-            dashboards_url='/dashboardmodelview/list'
-        )
-
-
-appbuilder.add_view(
-    DashboardModelView,
-    "Dashboards",
-    label=__("Dashboards"),
-    icon="fa-dashboard",
-    category='',
-    category_icon='',)
-
-
-class DashboardModelViewAsync(DashboardModelView):  # noqa
-    list_columns = ['dashboard_link', 'creator', 'modified', 'dashboard_title']
-    label_columns = {
-        'dashboard_link': _('Dashboard'),
-        'dashboard_title': _('Title'),
-        'creator': _('Creator'),
-        'modified': _('Modified'),
-    }
-
-appbuilder.add_view_no_menu(DashboardModelViewAsync)
 
 
 class LogModelView(SupersetModelView):
@@ -2237,7 +1887,6 @@ class CssTemplateModelView(SupersetModelView, DeleteMixin):
 class CssTemplateAsyncModelView(CssTemplateModelView):
     list_columns = ['template_name', 'css']
 
-appbuilder.add_separator("Sources")
 appbuilder.add_view(
     CssTemplateModelView,
     "CSS Templates",
@@ -2248,25 +1897,6 @@ appbuilder.add_view(
     category_icon='')
 
 appbuilder.add_view_no_menu(CssTemplateAsyncModelView)
-
-appbuilder.add_link(
-    'SQL Editor',
-    label=_("SQL Editor"),
-    href='/superset/sqllab',
-    category_icon="fa-flask",
-    icon="fa-flask",
-    category='SQL Lab',
-    category_label=__("SQL Lab"),
-)
-appbuilder.add_link(
-    'Query Search',
-    label=_("Query Search"),
-    href='/superset/sqllab#search',
-    icon="fa-search",
-    category_icon="fa-flask",
-    category='SQL Lab',
-    category_label=__("SQL Lab"),
-)
 
 
 @app.after_request
